Command execution. Not tested yet.
