--- osu-al1s-core/src/main/java/moe/orangemc/osu/al1s/chat/web/model/OutboundChannelMessage.java
@@ -1,56 +1,0 @@
-/*
- * Copyright 2024 Astro angelfish
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-package moe.orangemc.osu.al1s.chat.web.model;
-
-import com.google.gson.TypeAdapter;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-
-import java.io.IOException;
-
-public record OutboundChannelMessage(String message, boolean action) {
-    public static class Adapter extends TypeAdapter<OutboundChannelMessage> {
-        @Override
-        public void write(JsonWriter jsonWriter, OutboundChannelMessage outboundChannelMessage) throws IOException {
-            jsonWriter.beginObject();
-            jsonWriter.name("message").value(outboundChannelMessage.message());
-            jsonWriter.name("is_action").value(outboundChannelMessage.action());
-            jsonWriter.endObject();
-        }
-
-        @Override
-        public OutboundChannelMessage read(JsonReader jsonReader) throws IOException {
-            jsonReader.beginObject();
-            String message = null;
-            boolean action = false;
-            while (jsonReader.hasNext()) {
-                switch (jsonReader.nextName()) {
-                    case "message":
-                        message = jsonReader.nextString();
-                        break;
-                    case "is_action":
-                        action = jsonReader.nextBoolean();
-                        break;
-                    default:
-                        jsonReader.skipValue();
-                }
-            }
-            jsonReader.endObject();
-            return new OutboundChannelMessage(message, action);
-        }
-    }
-}
