+++ osu-al1s-core/src/main/java/moe/orangemc/osu/al1s/event/accessor/LineNumberedMethodVisitor.java
@@ -1,0 +1,135 @@
+/*
+ * Copyright 2024 Astro angelfish
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package moe.orangemc.osu.al1s.event.accessor;
+
+import org.objectweb.asm.*;
+
+public class LineNumberedMethodVisitor extends MethodVisitor {
+    public static boolean debug = false;
+
+    protected LineNumberedMethodVisitor(MethodVisitor methodVisitor) {
+        super(Opcodes.ASM9, methodVisitor);
+    }
+
+    private void markLineNumber() {
+        if (!debug) {
+            return;
+        }
+
+        Label label = new Label();
+        int lineNumber = Thread.currentThread().getStackTrace()[3].getLineNumber();
+        super.visitLabel(label);
+        super.visitLineNumber(lineNumber, label);
+    }
+
+    @Override
+    public void visitCode() {
+        super.visitCode();
+    }
+
+    @Override
+    public void visitInsn(int opcode) {
+        markLineNumber();
+        super.visitInsn(opcode);
+    }
+
+    @Override
+    public void visitIntInsn(int opcode, int operand) {
+        markLineNumber();
+        super.visitIntInsn(opcode, operand);
+    }
+
+    @Override
+    public void visitVarInsn(int opcode, int varIndex) {
+        markLineNumber();
+        super.visitVarInsn(opcode, varIndex);
+    }
+
+    @Override
+    public void visitTypeInsn(int opcode, String type) {
+        markLineNumber();
+        super.visitTypeInsn(opcode, type);
+    }
+
+    @Override
+    public void visitFieldInsn(int opcode, String owner, String name, String descriptor) {
+        markLineNumber();
+        super.visitFieldInsn(opcode, owner, name, descriptor);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Deprecated
+    @Override
+    public void visitMethodInsn(int opcode, String owner, String name, String descriptor) {
+        markLineNumber();
+        super.visitMethodInsn(opcode, owner, name, descriptor);
+    }
+
+    @Override
+    public void visitMethodInsn(int opcode, String owner, String name, String descriptor, boolean isInterface) {
+        markLineNumber();
+        super.visitMethodInsn(opcode, owner, name, descriptor, isInterface);
+    }
+
+    @Override
+    public void visitInvokeDynamicInsn(String name, String descriptor, Handle bootstrapMethodHandle, Object... bootstrapMethodArguments) {
+        markLineNumber();
+        super.visitInvokeDynamicInsn(name, descriptor, bootstrapMethodHandle, bootstrapMethodArguments);
+    }
+
+    @Override
+    public void visitJumpInsn(int opcode, Label label) {
+        markLineNumber();
+        super.visitJumpInsn(opcode, label);
+    }
+
+    @Override
+    public void visitLabel(Label label) {
+        markLineNumber();
+        super.visitLabel(label);
+    }
+
+    @Override
+    public void visitLdcInsn(Object value) {
+        markLineNumber();
+        super.visitLdcInsn(value);
+    }
+
+    @Override
+    public void visitIincInsn(int varIndex, int increment) {
+        markLineNumber();
+        super.visitIincInsn(varIndex, increment);
+    }
+
+    @Override
+    public void visitTableSwitchInsn(int min, int max, Label dflt, Label... labels) {
+        markLineNumber();
+        super.visitTableSwitchInsn(min, max, dflt, labels);
+    }
+
+    @Override
+    public void visitLookupSwitchInsn(Label dflt, int[] keys, Label[] labels) {
+        markLineNumber();
+        super.visitLookupSwitchInsn(dflt, keys, labels);
+    }
+
+    @Override
+    public void visitMultiANewArrayInsn(String descriptor, int numDimensions) {
+        markLineNumber();
+        super.visitMultiANewArrayInsn(descriptor, numDimensions);
+    }
+}
